[refactor] 사용성 향상 위해 RecyclerView Adapter -> ListAdapter로 변경
