--- 최창익/app/src/main/java/com/mash_up/mvvmstudy/view/MainAdapter.kt
@@ -1,34 +1,0 @@
-package com.mash_up.mvvmstudy.view
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-import com.mash_up.mvvmstudy.databinding.ItemListBinding
-import com.mash_up.mvvmstudy.model.Repository
-
-class MainAdapter : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
-    private val items: MutableList<Repository> = mutableListOf()
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder =
-        MainViewHolder(
-            ItemListBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        )
-
-    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
-        (holder as MainViewHolder).bind(items[position])
-    }
-
-    override fun getItemCount(): Int =
-        items.size
-
-    fun submitList(items: List<Repository>) {
-        this.items.addAll(items)
-        notifyDataSetChanged()
-    }
-
-    class MainViewHolder(val binding: ItemListBinding) : RecyclerView.ViewHolder(binding.root) {
-        fun bind(repository: Repository) {
-            binding.repository = repository
-        }
-    }
-}
