+++ 최창익/app/src/main/java/com/mash_up/mvvmstudy/view/RepositoryAdapter.kt
@@ -1,0 +1,35 @@
+package com.mash_up.mvvmstudy.view
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
+import com.mash_up.mvvmstudy.databinding.ItemListBinding
+import com.mash_up.mvvmstudy.model.Repository
+
+class RepositoryAdapter : ListAdapter<Repository, RecyclerView.ViewHolder>(RepositoryDiffUtil()) {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder =
+        RepositoryViewHolder(
+            ItemListBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        )
+
+    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
+        (holder as RepositoryViewHolder).bind(getItem(position))
+    }
+
+    class RepositoryViewHolder(val binding: ItemListBinding) :
+        RecyclerView.ViewHolder(binding.root) {
+        fun bind(repository: Repository) {
+            binding.repository = repository
+        }
+    }
+
+    class RepositoryDiffUtil : DiffUtil.ItemCallback<Repository>() {
+        override fun areItemsTheSame(oldItem: Repository, newItem: Repository): Boolean =
+            oldItem == newItem
+
+        override fun areContentsTheSame(oldItem: Repository, newItem: Repository): Boolean =
+            oldItem.id == newItem.id
+    }
+}
