+++ 최창익/app/src/main/res/layout/item_list.xml
@@ -17,8 +17,8 @@
 
         <ImageView
             android:id="@+id/iv_item"
-            android:layout_width="24dp"
-            android:layout_height="24dp"
+            android:layout_width="36dp"
+            android:layout_height="36dp"
             app:imageUrl="@{repository.owner.avatarUrl}"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
